+++ signaturescanner.h
@@ -12,7 +12,7 @@ class SignatureScanner
 public:
     struct module
     {
-        uintptr_t  Base, Size;
+        DWORD  Base, Size;
     };
 
     struct launcher
@@ -24,7 +24,7 @@ public:
                 const char* module = "GameAssembly.dll";
                 const char* pattern = "\xA1\x00\x00\x00\x00\xF6\x80\xBF\x00\x00\x00\x02\x74\x14\x83\x78\x70\x00\x75\x0E\x50\xE8\x00\x00\x00\x00\xA1\x00\x00\x00\x00\x83\xC4\x04\x8B\x40\x5C\x8B\x40\x1C\x85\xC0\x0F\x84\x00\x00\x00\x00\x3B\x70\x0C\x0F\x83";
                 const char* mask = "x????xxxxxxxxxxxxxxxxx????x????xxxxxxxxxxxxx????xxxxx";
-                uintptr_t address;
+                DWORD address;
             };
             PLH PLH;
 
@@ -33,7 +33,7 @@ public:
                 const char* module = "GameAssembly.dll";
                 const char* pattern = "\xA1\x00\x00\x00\x00\xF6\x80\xBF\x00\x00\x00\x02\x74\x0F\x83\x78\x70\x00\x75\x09\x50\xE8\x00\x00\x00\x00\x83\xC4\x04\xF3\x0F\x10\x45\xF0";
                 const char* mask = "x????xxxxxxxxxxxxxxxxx????xxxxxxxx";
-                uintptr_t address;
+                DWORD address;
             };
             VWIK VWIK;
 
@@ -42,7 +42,7 @@ public:
                 const char* module = "GameAssembly.dll";
                 const char* pattern = "\xA1\x00\x00\x00\x00\xF6\x80\xBF\x00\x00\x00\x02\x74\x14\x83\x78\x70\x00\x75\x0E\x50\xE8\x00\x00\x00\x00\xA1\x00\x00\x00\x00\x83\xC4\x04\x8B\x40\x5C\x56\x8B\x70\x08\xA1\x00\x00\x00\x00\xF6\x80\xBF\x00\x00\x00\x02\x74\x0F\x83\x78\x70\x00\x75\x09\x50\xE8\x00\x00\x00\x00\x83\xC4\x04\x6A\x00\x6A\x00\x56\xE8\x00\x00\x00\x00\x83\xC4\x0C\x5E\x84\xC0\x75\x72";
                 const char* mask = "x????xxxxxxxxxxxxxxxxx????x????xxxxxxxxxxx????xxxxxxxxxxxxxxxxx????xxxxxxxxx????xxxxxxxx";
-                uintptr_t address;
+                DWORD address;
             };
             Controll Controll;
 
@@ -51,7 +51,7 @@ public:
                 const char* module = "GameAssembly.dll";
                 const char* pattern = "\xA1\x00\x00\x00\x00\x8B\x40\x5C\xC7\x00\x00\x00\x00\x00\xA1\x00\x00\x00\x00\x8B\x40\x5C\xC6\x40\x04\x00\xA1\x00\x00\x00\x00\x8B\x40\x5C\xC6\x40\x05\x00";
                 const char* mask = "x????xxxxxxxxxx????xxxxxxxx????xxxxxxx";
-                uintptr_t address;
+                DWORD address;
             };
             MainManager MainManager;
 
@@ -60,7 +60,7 @@ public:
                 const char* module = "GameAssembly.dll";
                 const char* pattern = "\xA1\x00\x00\x00\x00\x8B\x40\x5C\xC6\x00\x00\xA1\x00\x00\x00\x00\x8B\x40\x5C\xC6\x40\x01\x00\xA1\x00\x00\x00\x00\x8B\x40\x5C\xC6\x40\x02\x00\xA1\x00\x00\x00\x00\xF6\x80\xBF\x00\x00\x00\x02";
                 const char* mask = "x????xxxxxxx????xxxxxxxx????xxxxxxxx????xxxxxxx";
-                uintptr_t address;
+                DWORD address;
             };
             GUIInv GUIInv;
 
@@ -75,7 +75,7 @@ public:
                 const char* module = "GameAssembly.dll";
                 const char* pattern = "\x55\x8B\xEC\x80\x3D\x00\x00\x00\x00\x00\x75\x15\xFF\x35\x00\x00\x00\x00\xE8\x00\x00\x00\x00\x83\xC4\x04\xC6\x05\x00\x00\x00\x00\x00\xA1\x00\x00\x00\x00\xF6\x80\xBF\x00\x00\x00\x02\x74\x14\x83\x78\x70\x00\x75\x0E\x50\xE8\x00\x00\x00\x00\xA1\x00\x00\x00\x00\x83\xC4\x04\x8B\x40\x5C\x56\x57\xFF\x35\x00\x00\x00\x00\x8B\x78\x48\xE8\x00\x00\x00\x00\xF3\x0F\x10\x45\x14\x8B\xF0\xF3\x0F\x58\x45\x18\x6A\x00\x56\xF3\x0F\x11\x45\x18\xE8\x00\x00\x00\x00\xF3\x0F\x7E\x45\x08\x83\xC4\x0C\xF3\x0F\x10\x4D\x18\x8B\x45\x10\x66\x0F\xD6\x46\x08\xF3\x0F\x10\x45\x14\xF3\x0F\x11\x4E\x24\xF3\x0F\x5C\xC8\x89\x46\x10\x8A\x45\x1C\xF3\x0F\x11\x46\x20\xC7\x46\x2C\x06\x00\x00\x00";
                 const char* mask = "xxxxx?????xxxx????x????xxxxx?????x????xxxxxxxxxxxxxxxxx????x????xxxxxxxxxx????xxxx????xxxxxxxxxxxxxxxxxxxxx????xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx";
-                uintptr_t address;
+                DWORD address;
             };
             CameraAddAngle CameraAddAngle;
 
@@ -85,7 +85,7 @@ public:
                 const char* module = "GameAssembly.dll";
                 const char* pattern = "\x55\x8B\xEC\x51\x80\x3D\x00\x00\x00\x00\x00\x75\x15\xFF\x35\x00\x00\x00\x00\xE8\x00\x00\x00\x00\x83\xC4\x04\xC6\x05\x00\x00\x00\x00\x00\xA1\x00\x00\x00\x00\xF6\x80\xBF\x00\x00\x00\x02\x74\x14\x83\x78\x70\x00\x75\x0E\x50\xE8\x00\x00\x00\x00\xA1\x00\x00\x00\x00\x83\xC4\x04\x8B\x40\x5C\x56\x6A\x00\xC6\x80\x28\x01\x00\x00\x01";
                 const char* mask = "xxxxxx?????xxxx????x????xxxxx?????x????xxxxxxxxxxxxxxxxx????x????xxxxxxxxxxxxxxxx";
-                uintptr_t address;
+                DWORD address;
             };
             SetFreeze SetFreeze;
 
@@ -95,7 +95,7 @@ public:
                 const char* module = "GameAssembly.dll";
                 const char* pattern = "\x55\x8B\xEC\x80\x3D\x00\x00\x00\x00\x00\x75\x15\xFF\x35\x00\x00\x00\x00\xE8\x00\x00\x00\x00\x83\xC4\x04\xC6\x05\x00\x00\x00\x00\x00\xA1\x00\x00\x00\x00\xF6\x80\xBF\x00\x00\x00\x02\x74\x0F\x83\x78\x70\x00\x75\x09\x50\xE8\x00\x00\x00\x00\x83\xC4\x04\x57\x6A\x00\x6A\x04";
                 const char* mask = "xxxxx?????xxxx????x????xxxxx?????x????xxxxxxxxxxxxxxxxx????xxxxxxxx";
-                uintptr_t address;
+                DWORD address;
             };
             SendAttack SendAttack;
 
@@ -105,7 +105,7 @@ public:
                 const char* module = "GameAssembly.dll";
                 const char* pattern = "\x55\x8B\xEC\x80\x3D\x00\x00\x00\x00\x00\x75\x15\xFF\x35\x00\x00\x00\x00\xE8\x00\x00\x00\x00\x83\xC4\x04\xC6\x05\x00\x00\x00\x00\x00\xA1\x00\x00\x00\x00\xF6\x80\xBF\x00\x00\x00\x02\x74\x14\x83\x78\x70\x00\x75\x0E\x50\xE8\x00\x00\x00\x00\xA1\x00\x00\x00\x00\x83\xC4\x04\x8B\x40\x5C\x56\x57\xFF\x35\x00\x00\x00\x00\x8B\x78\x48\xE8\x00\x00\x00\x00\xF3\x0F\x10\x45\x14\x8B\xF0\xF3\x0F\x58\x45\x18\x6A\x00\x56\xF3\x0F\x11\x45\x18\xE8\x00\x00\x00\x00\xF3\x0F\x7E\x45\x08\x83\xC4\x0C\xF3\x0F\x10\x4D\x18\x8B\x45\x10\x66\x0F\xD6\x46\x08\xF3\x0F\x10\x45\x14\xF3\x0F\x11\x4E\x24\xF3\x0F\x5C\xC8\x89\x46\x10\x8A\x45\x1C\xF3\x0F\x11\x46\x20\xC7\x46\x2C\x04\x00\x00\x00";
                 const char* mask = "xxxxx?????xxxx????x????xxxxx?????x????xxxxxxxxxxxxxxxxx????x????xxxxxxxxxx????xxxx????xxxxxxxxxxxxxxxxxxxxx????xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx";
-                uintptr_t address;
+                DWORD address;
             };
             CameraAddOffset CameraAddOffset;
 
@@ -120,7 +120,7 @@ public:
                 const char* module = "GameAssembly.dll";
                 const char* pattern = "\xC6\x80\xBC\x00\x00\x00\x01";
                 const char* mask = "xxxxxxx";
-                uintptr_t address;
+                DWORD address;
             };
             AutomaticWeapon AutomaticWeapon;
 
@@ -130,7 +130,7 @@ public:
                 const char* module = "GameAssembly.dll";
                 const char* pattern = "\xFF\x48\x28\x8B\x40\x28";
                 const char* mask = "xxxxxx";
-                uintptr_t address;
+                DWORD address;
             };
             InfiniteAmmo InfiniteAmmo;
 
@@ -140,7 +140,7 @@ public:
                 const char* module = "GameAssembly.dll";
                 const char* pattern = "\xC7\x04\x24\x00\x00\x40\x40\x8D\x4D\x94";
                 const char* mask = "xxxxxxxxxx";
-                uintptr_t address;
+                DWORD address;
             };
             RangeShovels RangeShovels;
 
@@ -158,7 +158,7 @@ public:
                 const char* module = "GameAssembly.dll";
                 const char* pattern = "\xA1\x00\x00\x00\x00\xF6\x80\xBF\x00\x00\x00\x02\x74\x14\x83\x78\x70\x00\x75\x0E\x50\xE8\x00\x00\x00\x00\xA1\x00\x00\x00\x00\x83\xC4\x04\x8B\x40\x5C\x8B\x40\x1C\x85\xC0\x0F\x84\x00\x00\x00\x00\x3B\x70\x0C\x0F\x83";
                 const char* mask = "x????xxxxxxxxxxxxxxxxx????x????xxxxxxxxxxxxx????xxxxx";
-                uintptr_t address;
+                DWORD address;
             };
             PLH PLH;
 
@@ -167,7 +167,7 @@ public:
                 const char* module = "GameAssembly.dll";
                 const char* pattern = "\xA1\x00\x00\x00\x00\xF6\x80\xBF\x00\x00\x00\x02\x74\x0F\x83\x78\x70\x00\x75\x09\x50\xE8\x00\x00\x00\x00\x83\xC4\x04\xF3\x0F\x10\x45\xF0";
                 const char* mask = "x????xxxxxxxxxxxxxxxxx????xxxxxxxx";
-                uintptr_t address;
+                DWORD address;
             };
             VWIK VWIK;
 
@@ -176,7 +176,7 @@ public:
                 const char* module = "GameAssembly.dll";
                 const char* pattern = "\xA1\x00\x00\x00\x00\xF6\x80\xBF\x00\x00\x00\x02\x74\x14\x83\x78\x70\x00\x75\x0E\x50\xE8\x00\x00\x00\x00\xA1\x00\x00\x00\x00\x83\xC4\x04\x8B\x40\x5C\x56\x8B\x70\x08\xA1\x00\x00\x00\x00\xF6\x80\xBF\x00\x00\x00\x02\x74\x0F\x83\x78\x70\x00\x75\x09\x50\xE8\x00\x00\x00\x00\x83\xC4\x04\x6A\x00\x6A\x00\x56\xE8\x00\x00\x00\x00\x83\xC4\x0C\x5E\x84\xC0\x75\x72";
                 const char* mask = "x????xxxxxxxxxxxxxxxxx????x????xxxxxxxxxxx????xxxxxxxxxxxxxxxxx????xxxxxxxxx????xxxxxxxx";
-                uintptr_t address;
+                DWORD address;
             };
             Controll Controll;
 
@@ -185,7 +185,7 @@ public:
                 const char* module = "GameAssembly.dll";
                 const char* pattern = "\xA1\x00\x00\x00\x00\x8B\x40\x5C\xC7\x00\x00\x00\x00\x00\xA1\x00\x00\x00\x00\x8B\x40\x5C\xC6\x40\x04\x00\xA1\x00\x00\x00\x00\x8B\x40\x5C\xC6\x40\x05\x00";
                 const char* mask = "x????xxxxxxxxxx????xxxxxxxx????xxxxxxx";
-                uintptr_t address;
+                DWORD address;
             };
             MainManager MainManager;
 
@@ -194,7 +194,7 @@ public:
                 const char* module = "GameAssembly.dll";
                 const char* pattern = "\xA1\x00\x00\x00\x00\x8B\x40\x5C\xC6\x00\x00\xA1\x00\x00\x00\x00\x8B\x40\x5C\xC6\x40\x01\x00\xA1\x00\x00\x00\x00\x8B\x40\x5C\xC6\x40\x02\x00\xA1\x00\x00\x00\x00\xF6\x80\xBF\x00\x00\x00\x02";
                 const char* mask = "x????xxxxxxx????xxxxxxxx????xxxxxxxx????xxxxxxx";
-                uintptr_t address;
+                DWORD address;
             };
             GUIInv GUIInv;
 
@@ -209,7 +209,7 @@ public:
                 const char* module = "GameAssembly.dll";
                 const char* pattern = "\x55\x8B\xEC\x80\x3D\x00\x00\x00\x00\x00\x75\x15\xFF\x35\x00\x00\x00\x00\xE8\x00\x00\x00\x00\x83\xC4\x04\xC6\x05\x00\x00\x00\x00\x00\xA1\x00\x00\x00\x00\xF6\x80\xBF\x00\x00\x00\x02\x74\x14\x83\x78\x70\x00\x75\x0E\x50\xE8\x00\x00\x00\x00\xA1\x00\x00\x00\x00\x83\xC4\x04\x8B\x40\x5C\x56\x57\xFF\x35\x00\x00\x00\x00\x8B\x78\x48\xE8\x00\x00\x00\x00\xF3\x0F\x10\x45\x14\x8B\xF0\xF3\x0F\x58\x45\x18\x6A\x00\x56\xF3\x0F\x11\x45\x18\xE8\x00\x00\x00\x00\xF3\x0F\x7E\x45\x08\x83\xC4\x0C\xF3\x0F\x10\x4D\x18\x8B\x45\x10\x66\x0F\xD6\x46\x08\xF3\x0F\x10\x45\x14\xF3\x0F\x11\x4E\x24\xF3\x0F\x5C\xC8\x89\x46\x10\x8A\x45\x1C\xF3\x0F\x11\x46\x20\xC7\x46\x2C\x06\x00\x00\x00";
                 const char* mask = "xxxxx?????xxxx????x????xxxxx?????x????xxxxxxxxxxxxxxxxx????x????xxxxxxxxxx????xxxx????xxxxxxxxxxxxxxxxxxxxx????xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx";
-                uintptr_t address;
+                DWORD address;
             };
             CameraAddAngle CameraAddAngle;
 
@@ -219,7 +219,7 @@ public:
                 const char* module = "GameAssembly.dll";
                 const char* pattern = "\x55\x8B\xEC\x51\x80\x3D\x00\x00\x00\x00\x00\x75\x15\xFF\x35\x00\x00\x00\x00\xE8\x00\x00\x00\x00\x83\xC4\x04\xC6\x05\x00\x00\x00\x00\x00\xA1\x00\x00\x00\x00\xF6\x80\xBF\x00\x00\x00\x02\x74\x14\x83\x78\x70\x00\x75\x0E\x50\xE8\x00\x00\x00\x00\xA1\x00\x00\x00\x00\x83\xC4\x04\x8B\x40\x5C\x56\x6A\x00\xC6\x80\x28\x01\x00\x00\x01";
                 const char* mask = "xxxxxx?????xxxx????x????xxxxx?????x????xxxxxxxxxxxxxxxxx????x????xxxxxxxxxxxxxxxx";
-                uintptr_t address;
+                DWORD address;
             };
             SetFreeze SetFreeze;
 
@@ -229,7 +229,7 @@ public:
                 const char* module = "GameAssembly.dll";
                 const char* pattern = "\x55\x8B\xEC\x80\x3D\x00\x00\x00\x00\x00\x75\x15\xFF\x35\x00\x00\x00\x00\xE8\x00\x00\x00\x00\x83\xC4\x04\xC6\x05\x00\x00\x00\x00\x00\xA1\x00\x00\x00\x00\xF6\x80\xBF\x00\x00\x00\x02\x74\x0F\x83\x78\x70\x00\x75\x09\x50\xE8\x00\x00\x00\x00\x83\xC4\x04\x57\x6A\x00\x6A\x04";
                 const char* mask = "xxxxx?????xxxx????x????xxxxx?????x????xxxxxxxxxxxxxxxxx????xxxxxxxx";
-                uintptr_t address;
+                DWORD address;
             };
             SendAttack SendAttack;
 
@@ -239,7 +239,7 @@ public:
                 const char* module = "GameAssembly.dll";
                 const char* pattern = "\x55\x8B\xEC\x80\x3D\x00\x00\x00\x00\x00\x75\x15\xFF\x35\x00\x00\x00\x00\xE8\x00\x00\x00\x00\x83\xC4\x04\xC6\x05\x00\x00\x00\x00\x00\xA1\x00\x00\x00\x00\xF6\x80\xBF\x00\x00\x00\x02\x74\x14\x83\x78\x70\x00\x75\x0E\x50\xE8\x00\x00\x00\x00\xA1\x00\x00\x00\x00\x83\xC4\x04\x8B\x40\x5C\x56\x57\xFF\x35\x00\x00\x00\x00\x8B\x78\x48\xE8\x00\x00\x00\x00\xF3\x0F\x10\x45\x14\x8B\xF0\xF3\x0F\x58\x45\x18\x6A\x00\x56\xF3\x0F\x11\x45\x18\xE8\x00\x00\x00\x00\xF3\x0F\x7E\x45\x08\x83\xC4\x0C\xF3\x0F\x10\x4D\x18\x8B\x45\x10\x66\x0F\xD6\x46\x08\xF3\x0F\x10\x45\x14\xF3\x0F\x11\x4E\x24\xF3\x0F\x5C\xC8\x89\x46\x10\x8A\x45\x1C\xF3\x0F\x11\x46\x20\xC7\x46\x2C\x04\x00\x00\x00";
                 const char* mask = "xxxxx?????xxxx????x????xxxxx?????x????xxxxxxxxxxxxxxxxx????x????xxxxxxxxxx????xxxx????xxxxxxxxxxxxxxxxxxxxx????xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx";
-                uintptr_t address;
+                DWORD address;
             };
             CameraAddOffset CameraAddOffset;
 
@@ -254,7 +254,7 @@ public:
                 const char* module = "GameAssembly.dll";
                 const char* pattern = "\xC6\x80\xBC\x00\x00\x00\x01";
                 const char* mask = "xxxxxxx";
-                uintptr_t address;
+                DWORD address;
             };
             AutomaticWeapon AutomaticWeapon;
 
@@ -264,7 +264,7 @@ public:
                 const char* module = "GameAssembly.dll";
                 const char* pattern = "\xFF\x48\x28\x8B\x40\x28";
                 const char* mask = "xxxxxx";
-                uintptr_t address;
+                DWORD address;
             };
             InfiniteAmmo InfiniteAmmo;
 
@@ -274,7 +274,7 @@ public:
                 const char* module = "GameAssembly.dll";
                 const char* pattern = "\xC7\x04\x24\x00\x00\x40\x40\x8D\x4D\x94";
                 const char* mask = "xxxxxxxxxx";
-                uintptr_t address;
+                DWORD address;
             };
             RangeShovels RangeShovels;
 
@@ -287,7 +287,7 @@ public:
 
     module TargetModule;
     HANDLE TargetProcess;
-    uintptr_t TargetId;
+    DWORD TargetId;
 
 };
 
@@ -303,7 +303,7 @@ namespace Scanner
     DWORD GetModuleSize(DWORD pID, char *moduleName);
 
     BOOL ComparePattern(HANDLE pHandle, DWORD address, char *pattern, char *mask);
-    uintptr_t ExternalAoBScan(HANDLE pHandle, uintptr_t pID, char *mod, char *pattern, char *mask);
+    DWORD ExternalAoBScan(HANDLE pHandle, DWORD pID, char *mod, char *pattern, char *mask);
 
 };
